aptly publish update to GCS now works
The ETag header no longer contains the MD5 hash, so we have to pull that
from a different field in the Object struct.  Warn about possible
composite objects, which don't have an MD5 hash, just a CRC32C checksum.

This incarnation leaves behind .tmp files for most files in the dists
directory; not sure why yet.